[sample] Renamed select to selectable and fixed square's selection
Fixed #164
