+++ sample/src/main/res/layout/activity_multi_selectable.xml
@@ -20,7 +20,7 @@
     android:id="@+id/activity_multi_grid"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="me.drakeet.multitype.sample.multi_select.MultiSelectActivity">
+    tools:context="me.drakeet.multitype.sample.multi_selectable.MultiSelectableActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/list"
